+++ docs/source/conf.py
@@ -75,7 +75,7 @@ extensions = [
     "sphinx.ext.viewcode",
     "sphinx_automodapi.automodapi",
     "sphinx_automodapi.smart_resolver",
-    # 'texext',
+    "texext",
 ]
